+++ site/src/components/playground-monaco.ts
@@ -1,0 +1,6 @@
+import * as monaco from 'monaco-editor'
+import EditorWorker from 'monaco-editor/esm/vs/editor/editor.worker?worker&inline'
+import TsWorker from 'monaco-editor/esm/vs/language/typescript/ts.worker?worker&inline'
+import CssWorker from 'monaco-editor/esm/vs/language/css/css.worker?worker&inline'
+
+export { monaco, EditorWorker, TsWorker, CssWorker }